# LikeFood/logic/backend/swagger_server/models/categories.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from logic.backend.swagger_server.models.base_model_ import Model
from logic.backend.swagger_server.models.category import Category  # noqa: F401,E501
from logic.backend.swagger_server import util


class Categories(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self):  # noqa: E501
        """Categories - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'Categories':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Categories of this Categories.  # noqa: E501
        :rtype: Categories
        """
        return util.deserialize_model(dikt, cls)