# LikeFood/logic/backend/swagger_server/models/category.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from logic.backend.swagger_server.models.base_model_ import Model
from logic.backend.swagger_server import util


class Category(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, title: str=None):  # noqa: E501
        """Category - a model defined in Swagger

        :param id: The id of this Category.  # noqa: E501
        :type id: int
        :param title: The title of this Category.  # noqa: E501
        :type title: str
        """
        self.swagger_types = {
            'id': int,
            'title': str
        }

        self.attribute_map = {
            'id': 'id',
            'title': 'title'
        }
        self._id = id
        self._title = title

    @classmethod
    def from_dict(cls, dikt) -> 'Category':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Category of this Category.  # noqa: E501
        :rtype: Category
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Category.


        :return: The id of this Category.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Category.


        :param id: The id of this Category.
        :type id: int
        """

        self._id = id

    @property
    def title(self) -> str:
        """Gets the title of this Category.


        :return: The title of this Category.
        :rtype: str
        """
        return self._title

    @title.setter
    def title(self, title: str):
        """Sets the title of this Category.


        :param title: The title of this Category.
        :type title: str
        """
        allowed_values = ["Первые блюда", "Вторые блюда", "Салаты", "Закуски", "Десерты"]  # noqa: E501
        if title not in allowed_values:
            raise ValueError(
                "Invalid value for `title` ({0}), must be one of {1}"
                .format(title, allowed_values)
            )

        self._title = title